trade_portal/requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.0.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
redis==3.4.1 # https://github.com/andymccurdy/redis-py
celery==4.4.0  # pyup: < 5.0  # https://github.com/celery/celery
flower==0.9.3  # https://github.com/mher/flower
psycopg2==2.8.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
pycountry==19.8.18

# Django
# ------------------------------------------------------------------------------
django==2.2.10  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/niwinz/django-redis
whitenoise==5.0.1
dj-pagination==2.4.0
django-widget-tweaks==1.4.8
django-constance[database]==2.6.0
django-ratelimit==2.0.0
django-recaptcha==2.0.6

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

sentry-sdk==0.14.1  # https://github.com/getsentry/sentry-python
django-storages[boto3]==1.9.1  # https://github.com/jschneier/django-storages

# For KMS / env decryption
boto3>=1.7.72,<2
awscli>=1.15.76,<2

# AWS Cognito Auth
mozilla-django-oidc==1.2.3
